sideride_flask_backend/requirements.txt
aniso8601==9.0.1
click==8.0.3
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mysql-connector-python==8.0.27
protobuf==3.19.1
pytz==2021.3
six==1.16.0
